requirements.txt
flask
groq
python-dotenv
requests
gunicorn